Testing floating error.

+++ src/protocol.c
@@ -128,9 +128,7 @@ unsigned char hdp_recieve(void) {
     if (DEBUG) { platform_debug("Bitrate", (1000 / temp_delay)); }
 
     temp_delay = (1000 / (float)temp_delay);
-
-    //_recieve_delay_per_bit = temp_delay;
-    _recieve_delay_per_bit = (1000 / temp_delay);
+    _recieve_delay_per_bit = (int)(1000 / temp_delay);
 
     if (DEBUG) { platform_debug("Delay per bit", _recieve_delay_per_bit); }
     if (DEBUG) { platform_debug("Bitrate", (1000 / _recieve_delay_per_bit)); }
